Wrote compressor functions to start from a specified level structure
The idea is that if you compress the database in chunks, you will get
some inneficiencies (especially with small chunks) as when starting
from scratch again, you have to create a new snapshot (with many table
rows!). If you start from the level structure that you ended the last
chunk with, it is exactly as though you just kept running - allowing
you to run the compressor in as small batches as you want!