Switch docker cache to github actions backend
We were failing to push cached images to the registry, because we're no
longer logging to the GHCR. In addition, secrets.GITHUB_TOKEN would need
to be expanded to support writing to the container registry, and giving
that to arbitrary PRs isn't great.

The GitHub Actions build cache is auto-clearing after 7 days, and
requires no auth.